Remove bootstrap, its no longer required as Gradle has a nicer error message for outdated java.
