--- bootstrap/.gitignore
@@ -1,7 +1,0 @@
-# Ignore everything
-/*
-
-!/src
-!/build.gradle
-!/.gitignore
-!/test-project
