--- bootstrap/build.gradle
@@ -1,32 +1,0 @@
-plugins {
-	id 'java'
-	id 'groovy'
-}
-
-java {
-	sourceCompatibility = JavaVersion.VERSION_1_8
-	targetCompatibility = JavaVersion.VERSION_1_8
-}
-
-tasks.withType(JavaCompile).configureEach {
-	it.options.encoding = "UTF-8"
-	it.options.release = 8
-}
-
-repositories {
-	mavenCentral()
-}
-
-dependencies {
-	implementation gradleApi()
-
-	testImplementation(gradleTestKit())
-	testImplementation('org.spockframework:spock-core:2.3-groovy-3.0') {
-		exclude module: 'groovy-all'
-	}
-}
-
-test {
-	maxHeapSize = "4096m"
-	useJUnitPlatform()
-}
