--- bootstrap/src/main/java/net/fabricmc/loom/bootstrap/BootstrappedPlugin.java
@@ -1,7 +1,0 @@
-package net.fabricmc.loom.bootstrap;
-
-import org.gradle.api.plugins.PluginAware;
-
-public interface BootstrappedPlugin {
-	void apply(PluginAware project);
-}
