--- bootstrap/test-project/build.gradle
@@ -1,8 +1,0 @@
-plugins {
-	id 'fabric-loom' version '0.13.local'
-}
-
-dependencies {
-	minecraft "com.mojang:minecraft:1.16.5"
-	mappings loom.officialMojangMappings()
-}
