--- bootstrap/test-project/settings.gradle
@@ -1,10 +1,0 @@
-pluginManagement {
-    repositories {
-        maven {
-            name = 'Fabric'
-            url = 'https://maven.fabricmc.net/'
-        }
-        gradlePluginPortal()
-		mavenLocal()
-    }
-}
