+++ settings.gradle
@@ -10,5 +10,3 @@ dependencyResolutionManagement {
 		}
 	}
 }
-
-include "bootstrap"
